requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
pydantic==1.10.7
sentence-transformers==2.2.2
faiss-cpu==1.7.4
PyMuPDF==1.22.3
python-docx==0.8.11
numpy==1.24.3
python-dotenv==1.0.0
transformers==4.30.2
accelerate==0.20.3
bitsandbytes==0.39.0